fix: remove duplicate hook for permit2 approvals
