+++ src/nft/components/bag/BagFooter.test.tsx
@@ -5,13 +5,12 @@ import { useWeb3React } from '@web3-react/core'
 import { nativeOnChain } from 'constants/tokens'
 import { useNftUniversalRouterAddress } from 'graphql/data/nft/NftUniversalRouterAddress'
 import { useCurrency } from 'hooks/Tokens'
-import { AllowanceState } from 'hooks/usePermit2Allowance'
+import usePermit2Allowance, { AllowanceState } from 'hooks/usePermit2Allowance'
 import { useTokenBalance } from 'lib/hooks/useCurrencyBalance'
 import { useBag, useWalletBalance } from 'nft/hooks'
 import { useBagTotalEthPrice } from 'nft/hooks/useBagTotalEthPrice'
 import useDerivedPayWithAnyTokenSwapInfo from 'nft/hooks/useDerivedPayWithAnyTokenSwapInfo'
 import usePayWithAnyTokenSwap from 'nft/hooks/usePayWithAnyTokenSwap'
-import usePermit2Approval from 'nft/hooks/usePermit2Approval'
 import { usePriceImpact } from 'nft/hooks/usePriceImpact'
 import { useTokenInput } from 'nft/hooks/useTokenInput'
 import { BagStatus } from 'nft/types'
@@ -31,7 +30,7 @@ jest.mock('hooks/Tokens')
 jest.mock('nft/hooks/usePayWithAnyTokenSwap')
 jest.mock('nft/hooks/useDerivedPayWithAnyTokenSwapInfo')
 jest.mock('graphql/data/nft/NftUniversalRouterAddress')
-jest.mock('nft/hooks/usePermit2Approval')
+jest.mock('hooks/usePermit2Allowance')
 jest.mock('nft/hooks/usePriceImpact')
 
 const renderBagFooter = () => {
@@ -63,14 +62,9 @@ describe('BagFooter.tsx', () => {
       provider: undefined,
     })
 
-    mocked(usePermit2Approval).mockReturnValue({
-      allowance: {
-        state: AllowanceState.ALLOWED,
-        permitSignature: undefined,
-      },
-      isApprovalLoading: false,
-      isAllowancePending: false,
-      updateAllowance: () => Promise.resolve(),
+    mocked(usePermit2Allowance).mockReturnValue({
+      state: AllowanceState.ALLOWED,
+      permitSignature: undefined,
     })
     mocked(useNftUniversalRouterAddress).mockReturnValue({
       universalRouterAddress: undefined,
@@ -321,13 +315,8 @@ describe('BagFooter.tsx', () => {
   })
 
   it('loading allowance', () => {
-    mocked(usePermit2Approval).mockReturnValue({
-      allowance: {
-        state: AllowanceState.LOADING,
-      },
-      isAllowancePending: false,
-      isApprovalLoading: false,
-      updateAllowance: () => Promise.resolve(),
+    mocked(usePermit2Allowance).mockReturnValue({
+      state: AllowanceState.LOADING,
     })
 
     renderBagFooter()
@@ -338,24 +327,19 @@ describe('BagFooter.tsx', () => {
   })
 
   it('approval is loading', () => {
-    mocked(usePermit2Approval).mockReturnValue({
-      allowance: {
-        state: AllowanceState.REQUIRED,
-        isApprovalLoading: false,
-        isApprovalPending: false,
-        token: TEST_TOKEN_1,
-        approveAndPermit: () => Promise.resolve(),
-        approve: () => Promise.resolve(),
-        permit: () => Promise.resolve(),
-        revoke: () => Promise.resolve(),
-        needsSetupApproval: false,
-        needsPermitSignature: false,
-        isRevocationPending: false,
-        allowedAmount: CurrencyAmount.fromRawAmount(TEST_TOKEN_1, 0),
-      },
-      isAllowancePending: false,
+    mocked(usePermit2Allowance).mockReturnValue({
+      state: AllowanceState.REQUIRED,
       isApprovalLoading: true,
-      updateAllowance: () => Promise.resolve(),
+      isApprovalPending: false,
+      token: TEST_TOKEN_1,
+      approveAndPermit: () => Promise.resolve(),
+      approve: () => Promise.resolve(),
+      permit: () => Promise.resolve(),
+      revoke: () => Promise.resolve(),
+      needsSetupApproval: false,
+      needsPermitSignature: false,
+      isRevocationPending: false,
+      allowedAmount: CurrencyAmount.fromRawAmount(TEST_TOKEN_1, 0),
     })
 
     renderBagFooter()
@@ -366,24 +350,19 @@ describe('BagFooter.tsx', () => {
   })
 
   it('allowance to be confirmed in wallet', () => {
-    mocked(usePermit2Approval).mockReturnValue({
-      allowance: {
-        state: AllowanceState.REQUIRED,
-        isApprovalLoading: false,
-        isApprovalPending: false,
-        token: TEST_TOKEN_1,
-        approveAndPermit: () => Promise.resolve(),
-        approve: () => Promise.resolve(),
-        permit: () => Promise.resolve(),
-        revoke: () => Promise.resolve(),
-        needsSetupApproval: false,
-        needsPermitSignature: false,
-        isRevocationPending: false,
-        allowedAmount: CurrencyAmount.fromRawAmount(TEST_TOKEN_1, 0),
-      },
-      isAllowancePending: true,
+    mocked(usePermit2Allowance).mockReturnValue({
+      state: AllowanceState.REQUIRED,
       isApprovalLoading: false,
-      updateAllowance: () => Promise.resolve(),
+      isApprovalPending: true,
+      token: TEST_TOKEN_1,
+      approveAndPermit: () => Promise.resolve(),
+      approve: () => Promise.resolve(),
+      permit: () => Promise.resolve(),
+      revoke: () => Promise.resolve(),
+      needsSetupApproval: false,
+      needsPermitSignature: false,
+      isRevocationPending: false,
+      allowedAmount: CurrencyAmount.fromRawAmount(TEST_TOKEN_1, 0),
     })
 
     renderBagFooter()
@@ -394,24 +373,19 @@ describe('BagFooter.tsx', () => {
   })
 
   it('approve', () => {
-    mocked(usePermit2Approval).mockReturnValue({
-      allowance: {
-        state: AllowanceState.REQUIRED,
-        isApprovalLoading: false,
-        isApprovalPending: false,
-        token: TEST_TOKEN_1,
-        approveAndPermit: () => Promise.resolve(),
-        approve: () => Promise.resolve(),
-        permit: () => Promise.resolve(),
-        revoke: () => Promise.resolve(),
-        needsSetupApproval: false,
-        needsPermitSignature: false,
-        isRevocationPending: false,
-        allowedAmount: CurrencyAmount.fromRawAmount(TEST_TOKEN_1, 0),
-      },
-      isAllowancePending: false,
+    mocked(usePermit2Allowance).mockReturnValue({
+      state: AllowanceState.REQUIRED,
       isApprovalLoading: false,
-      updateAllowance: () => Promise.resolve(),
+      isApprovalPending: false,
+      token: TEST_TOKEN_1,
+      approveAndPermit: () => Promise.resolve(),
+      approve: () => Promise.resolve(),
+      permit: () => Promise.resolve(),
+      revoke: () => Promise.resolve(),
+      needsSetupApproval: false,
+      needsPermitSignature: false,
+      isRevocationPending: false,
+      allowedAmount: CurrencyAmount.fromRawAmount(TEST_TOKEN_1, 0),
     })
 
     renderBagFooter()
